Added AbstractFooterTest Updated inventory_tests.xml and positive_tests.xml
