+++ src/test/java/test_suites/inventory_tests.xml
@@ -5,6 +5,7 @@
         <classes>
             <class name="tests.inventory_tests.inventory_test.InventoryTest"/>
             <class name="tests.inventory_tests.sort_test.InventorySortTest"/>
+            <class name="tests.inventory_tests.footer_test.InventoryFooterTest"/>
         </classes>
     </test>
 </suite>
