+++ src/test/java/tests/abstract_tests/DriverType.java
@@ -1,4 +1,4 @@
-package tests.abstract_test;
+package tests.abstract_tests;
 
 public enum DriverType {
     CHROME,
