+++ src/test/java/tests/login_tests/login_test/LoginPreTest.java
@@ -2,7 +2,7 @@ package tests.login_tests.login_test;
 
 import org.testng.annotations.BeforeMethod;
 import pages.login_page.LoginPage;
-import tests.abstract_test.AbstractTest;
+import tests.abstract_tests.AbstractTest;
 
 public class LoginPreTest extends AbstractTest {
     protected LoginPage loginPage;
